removing some locations and moving tests files
