+++ apps/hubspot/package.json
