--- apps/hubspot/src/locations/EntryEditor.spec.tsx
@@ -1,17 +1,0 @@
-import EntryEditor from './EntryEditor';
-import { render } from '@testing-library/react';
-import { mockCma, mockSdk } from '../../test/mocks';
-import { vi } from 'vitest';
-
-vi.mock('@contentful/react-apps-toolkit', () => ({
-  useSDK: () => mockSdk,
-  useCMA: () => mockCma,
-}));
-
-describe('Entry component', () => {
-  it('Component text exists', () => {
-    const { getByText } = render(<EntryEditor />);
-
-    expect(getByText('Hello Entry Editor Component (AppId: test-app)')).toBeInTheDocument();
-  });
-});
