--- apps/hubspot/src/locations/Field.spec.tsx
@@ -1,17 +1,0 @@
-import Field from './Field';
-import { render } from '@testing-library/react';
-import { mockCma, mockSdk } from '../../test/mocks';
-import { vi } from 'vitest';
-
-vi.mock('@contentful/react-apps-toolkit', () => ({
-  useSDK: () => mockSdk,
-  useCMA: () => mockCma,
-}));
-
-describe('Field component', () => {
-  it('Component text exists', () => {
-    const { getByText } = render(<Field />);
-
-    expect(getByText('Hello Entry Field Component (AppId: test-app)')).toBeInTheDocument();
-  });
-});
