--- apps/hubspot/src/locations/Home.spec.tsx
@@ -1,17 +1,0 @@
-import Home from './Home';
-import { render } from '@testing-library/react';
-import { mockCma, mockSdk } from '../../test/mocks';
-import { vi } from 'vitest';
-
-vi.mock('@contentful/react-apps-toolkit', () => ({
-  useSDK: () => mockSdk,
-  useCMA: () => mockCma,
-}));
-
-describe('Home component', () => {
-  it('Component text exists', () => {
-    const { getByText } = render(<Home />);
-
-    expect(getByText('Hello Home Component (AppId: test-app)')).toBeInTheDocument();
-  });
-});
