+++ apps/hubspot/test/locations/ConfigScreen.spec.tsx
@@ -1,6 +1,6 @@
-import ConfigScreen from './ConfigScreen';
+import ConfigScreen from '../../src/locations/ConfigScreen';
 import { render } from '@testing-library/react';
-import { mockCma, mockSdk } from '../../test/mocks';
+import { mockCma, mockSdk } from '../mocks';
 import { vi } from 'vitest';
 
 vi.mock('@contentful/react-apps-toolkit', () => ({
