+++ apps/hubspot/test/locations/Page.spec.tsx
@@ -1,6 +1,6 @@
-import Page from './Page';
+import Page from '../../src/locations/Page';
 import { render } from '@testing-library/react';
-import { mockCma, mockSdk } from '../../test/mocks';
+import { mockCma, mockSdk } from '../mocks';
 import { vi } from 'vitest';
 
 vi.mock('@contentful/react-apps-toolkit', () => ({
