+++ apps/hubspot/test/locations/Sidebar.spec.tsx
@@ -1,6 +1,6 @@
-import Sidebar from './Sidebar';
+import Sidebar from '../../src/locations/Sidebar';
 import { render } from '@testing-library/react';
-import { mockCma, mockSdk } from '../../test/mocks';
+import { mockCma, mockSdk } from '../mocks';
 import { vi } from 'vitest';
 
 vi.mock('@contentful/react-apps-toolkit', () => ({
